requirements.txt
requests==2.31.0
pytest==8.1.1
selenium==4.19.0
pydantic-settings==2.2.1
PyYAML==6.0.1
allure-pytest==2.13.5
